apps/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
pydantic
supabase
python-dotenv
pydantic[email]
